Move StatementListProvider into the Statement namespace
