+++ Aliases.php
@@ -29,3 +29,12 @@ namespace Wikibase\DataModel\Claim {
 	class ClaimGuidParsingException extends \Wikibase\DataModel\Statement\StatementGuidParsingException {}
 
 }
+
+namespace Wikibase\DataModel {
+
+	/**
+	 * @deprecated since 3.0.0, use the base class instead.
+	 */
+	class StatementListProvider extends \Wikibase\DataModel\Statement\StatementListProvider {}
+
+}
